integrate encodings with code base
