+++ src/common/buffer/CellData.ts
@@ -4,7 +4,7 @@
  */
 
 import { CharData, ICellData } from 'common/Types';
-import { stringFromCodePoint } from 'common/input/TextDecoder';
+import { stringFromCodePoint } from 'common/input/Encodings';
 import { CHAR_DATA_CHAR_INDEX, CHAR_DATA_WIDTH_INDEX, CHAR_DATA_ATTR_INDEX, Content } from 'common/buffer/Constants';
 import { AttributeData } from 'common/buffer/AttributeData';
